Hoaithanh DONE write comic book

+++ Comic_app/app/src/main/java/com/example/comic_app/Fragments/Comic_Introduction_Fragment_Activity.java
@@ -154,7 +154,7 @@ public class Comic_Introduction_Fragment_Activity extends Fragment {
                         comicBook.setId(document.getId());
                         comicBook.setChapterList((List<String>) document.get("chapterList"));
                         comicBook.setAuthor((String)document.get("author"));
-                        comicBook.setCategory((List<Integer>)document.get("category"));
+                        comicBook.setCategory((List<Long>)document.get("category"));
                         comicBook.setImage((String)document.get("image"));
                         comicBook.setLength((String)document.get("length"));
                         comicBook.setStatus((String)document.get("status"));
